requirements.txt
google-generativeai>=0.3.2
google-api-python-client
chromadb
fastapi
uvicorn[standard]
readability-lxml
aiohttp
python-dotenv
